% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCisTrans.R
\name{assignCisTrans}
\alias{assignCisTrans}
\title{Assign eQTL as cis/trans}
\usage{
assignCisTrans(s1.output, perms.output, gff.cm, useConfInt = TRUE,
  cmInterval = 10, ...)
}
\arguments{
\item{s1.output}{Scanone output to be passed to calcCis}

\item{perms.output}{Permutation output to be passed to calcCis}

\item{gff.cm}{The gff with mapping position, as calculated from findGenecM}

\item{useConfInt}{Logical, should the confidence intervals be used to assign
cis/trans to eQTL?}

\item{cmInterval}{If useConfInt = FALSE, the cM distance the eQTL peak
needs to be from the physical position to constitute a cis-eQTL?}

\item{additional}{parameters passed to calcCis}
}
\description{
\code{assignCisTrans} An extension of qtlpvl::get.trans.info, but permitting
assignment using confidence intervals.
}

