% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCis.R
\name{calcCis}
\alias{calcCis}
\title{Calculate all confidence intervals from a QTL model}
\usage{
calcCis(mod, qtlnames = NULL, lodint = TRUE, drop = 1.5, prob = 0.95,
  ...)
}
\arguments{
\item{mod}{A QTL model containing LOD profiles. Can be generated by
stepwiseqtl (with keepLodProfile = TRUE), or by refineqtl.}

\item{qtlnames}{An optional vector of user-specified QTL identifiers}

\item{lodint}{Should lodint (default) or bayesint be used to estimate confidence intervals?}

\item{drop}{If lodint = TRUE, what drop threshold should be used?}

\item{prob}{If lodint = FALSE, what significance should be used for bayesint?}

\item{...}{Additional arguments passed on to either lodint or bayesint}
}
\value{
A dataframe of confidence intervals, including the maximum LOD score,
bounding markers and positions.
}
\description{
\code{calcCis} Takes a multiple QTL model and returns a dataframe of confidence intervals.
}
\details{
This function iterates through the qtl in the model, returning confidence intervals.
}
\examples{
library(qtlTools)
data(fake.bc)
cross<-fake.bc
cross <- calc.genoprob(cross, step=2.5)

# Make a QTL object and formula
mod <- makeqtl(cross, chr = c(2,5), pos = c(40,25), what = "prob")
sex <- data.frame(sex = cross$phe$sex)
nform <- "y ~ Q1 + Q2 + Q1*sex + sex"

#Calculate lodprofiles for confidence interval estimation
mod <- refineqtl(cross, mod, pheno.col = "pheno1",
  qtl = mod, formula = nform, covar = sex, method="hk")
calcCis(mod)

}

