% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQtlMeans.R
\name{calcQtlMeans}
\alias{calcQtlMeans}
\title{Using a QTL model, calcuate conditional QTL genotype effects}
\usage{
calcQtlMeans(cross, mod, covar = NULL, phe, form, ...)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{mod}{A QTL model, generated by either makeqtl (etc.) or stepwiseqtl.}

\item{phe}{Character or numeric vector indicating the phenotype to be tested}

\item{form}{If a generic (not stepwiseqtl) qtl model is supplied and does
not contain a formula, this is required. The formula must match the QTL in the
fitted QTL (mod) object.}
}
\value{
A dataframe of statistics.
}
\description{
\code{calcQtlMeans} Takes a QTL model, formula
}
\details{
This function works for riself, bc, f2 and 4-way mapping populations. It is advisable
to run both sim.geno and calc.genoprob prior to using qtlStats, as this will speed up the analysis.
Also, it is important to note that this function works best on models generated from stepwiseqtl.
}

