% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropByDropone.R
\name{dropByDropone}
\alias{dropByDropone}
\title{Drop markers that significantly expand the genetic map.}
\usage{
dropByDropone(cross, droponeRes, endMarkerThresh = 12,
  midMarkerThresh = 4, which.map = NULL, map.function = "kosambi",
  re.est.map = F, sex.sp = F, ...)
}
\arguments{
\item{cross}{The QTL cross object, with an estimated genetic map}

\item{droponeRes}{Results from qtl::droponemarker}

\item{endMarkerThresh}{The ldiff (length increase of map) that causes a marker
to be dropped from the map if the marker is at the end/start of a chromosome.}

\item{midMarkerThresh}{The ldiff (length increase of map) that causes a marker
to be dropped from the map.}

\item{which.map}{If the cross is a 4-way, which sex-specific map should be used?
See the column names of the dropone results. Defaults to Ldiff.female, but can be
set to Ldiff.male. Ignored if cross is not a 4way.}

\item{map.function}{If re.est.map = TRUE, the map functions to employ}

\item{re.est.map}{Should the genetic map be re-estimated?}

\item{sex.sp}{If re.est.map = TRUE, should a sex-specific map be estimated?}

\item{...}{Additional arguemnts passed on to est.map if re.est.map = TRUE.}
}
\value{
A cross object without markers that caused map expansion > thresholds.
}
\description{
\code{dropByDropone} Leverage the results of qtl::droponemarker to find markers
that are likely misplaced or erroneous.
}
\details{
A simple function to parse output from droponemarker postion.
}
\examples{
library(qtlTools)
data(fake.f2)
cross<-fake.f2
\dontrun{
dropone<-droponemarker(cross, map.function = "kosambi")
plot(dropone, lodcolumn = 2)

cross1<-dropByDropone(cross = cross, droponeRes = dropone, endMarkerThresh = 50, re.est.map=T)
plot.map(cross,cross1)
}
}
