% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCloseQTL.R
\name{dropCloseQTL}
\alias{dropCloseQTL}
\title{Remove close QTL from stepwiseqtl model selection}
\usage{
dropCloseQTL(cross, qtl, formula, covar, pheno.col, method = "hk",
  min.dist = 30)
}
\description{
\code{dropCloseQTL}  Assists in simplifying models following stepwise model selection
}
