% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropSmallCovInt.R
\name{dropSmallCovInt}
\alias{dropSmallCovInt}
\title{Reduce covariate interactions following stepwiseqtl.int}
\usage{
dropSmallCovInt(cross, qtl, formula, covar, pheno.col, method = "hk",
  thresh.diff)
}
\description{
\code{dropSmallCovInt}  Drops non-significant QTL*covariate interactions
}

