% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropTooCloseQTL.R
\name{dropTooCloseQTL}
\alias{dropTooCloseQTL}
\title{Remove close QTL from stepwiseqtl model selection}
\usage{
dropTooCloseQTL(cross, qtl, formula, covar, pheno.col, method = "hk",
  min.dist = 30)
}
\description{
\code{dropTooCloseQTL}  Assists in simplifying models following stepwise model selection
}
