% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGapsInMap.R
\name{fillGapsInMap}
\alias{fillGapsInMap}
\title{Find markers that have inferred positions within gaps in a map}
\usage{
fillGapsInMap(cross, minGapSize = 1, newMarGap = 1, marker.names,
  marker.chr, marker.pos, marker.lod, return4newLG = F, which.position = 3,
  ...)
}
\arguments{
\item{cross}{The QTL cross object.}

\item{minGapSize}{The smallest gap size to check for markers. Gaps in the map
smaller than this will be ignored.}

\item{marker.names}{The names of markers to check}

\item{marker.chr}{The inferred chromsomes of the markers - must match chrnames of
the cross.}

\item{marker.pos}{The inferred cM positions of the markers.}

\item{return4newLG}{Logical, Should the gaps be combined with the genetic map to pipe
straight into the newLG function?}

\item{...}{Not currently in use.}
}
\value{
If return4newLG = TRUE, a named and ordered list of all markers, including
those in the map. This output can be input directly into the newLG argument 'markerList'.
If return4newLG = FALSE, returns a simple named list of the markers that could be added.
List element names match the chromosomes that they belong to.
}
\description{
\code{fillGapsInMap} Using known cM positions and chromosomes,
determine which markers could be used to fill garps in a map.
}
\details{
A simple function to parse output from inferMarker postion
}
\examples{
library(qtlTools)
data(fake.bc)
cross<-fake.bc
\dontrun{
... more here ...
}
}
