% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.ciseqtl.R
\name{find.ciseqtl}
\alias{find.ciseqtl}
\title{Find the location of a cis eQTL}
\usage{
find.ciseqtl(cross, phe, pos.gene, chr.gene, cmWindow = 10,
  lodThreshold = 3, outputType = "data.frame", runTrans = TRUE,
  covar = NULL, ...)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{phe}{Character or numeric vector indicating the phenotype to be tested}

\item{pos.gene}{The suspected mapping position of the gene - may be inferred from bp2cm}

\item{chr.gene}{The name of the chromosome that contains the focal genes}

\item{cmWindow}{The distance from pos.gene that the cis eQTL may reside. Default is 10, but
if the confidence of the mapping position is high, a smaller value is suggested.}

\item{lodThreshold}{If outputType = data.frame, the presence of a cis.eqtl is defined by whether
any marker within cmWindow has a lod score > lodThreshold.}

\item{outputType}{Whether to return a data.frame or qtl object (from makeqtl)}

\item{lodThreshold}{}
}
\value{
Either a qtl object or a dataframe, indicating the position of a cis eQTL.
}
\description{
\code{find.ciseqtl} Find the mapping position of the highest lod score
within a window surrounding the known position of a gene.
}

