% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGenesInterval.R
\name{findGenesInterval}
\alias{findGenesInterval}
\title{Find candidate genes for QTL.}
\usage{
findGenesInterval(findGenecM.output = NULL, calcCis.output = NULL,
  qtlname = NULL, chr, lowposition, highposition, geneID, geneChr, geneBp,
  genecM)
}
\arguments{
\item{findGenecM.output}{Output from findGenecM
If supplied, geneID, geneChr, geneBp and genecM are ignored}

\item{calcCis.output}{Output from calcCis.
If supplied, chr, lowposition and highposition are ignored}

\item{qtlname}{Identifier of each qtl}

\item{chr}{QTL Chromosome}

\item{lowposition}{Lower confidence interval bound (cM)}

\item{highposition}{Upper confidence interval bound (cM)}

\item{geneID}{The name of each gene}

\item{geneChr}{The chromosome of each gene}

\item{geneBp}{The basepair position of each gene (not currently in use)}

\item{genecM}{The mapping position of each gene}

\item{...}{Not currently in use.}
}
\value{
A named list where each element corresponds to a provided QTL
confidence interval. Character vectors in each element display potential
candidate genes for each QTL.
}
\description{
\code{findGenesInterval} Use the mapping position (inferred) of each gene to
find potential candidates under a QTL confidence interval
}

