% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSimilarMarkers.R
\name{findSimilarMarkers}
\alias{findSimilarMarkers}
\title{Method to improve a genetic map.}
\usage{
findSimilarMarkers(cross, chr = NULL, rf.threshold = 0.02, sd.weight = 1,
  na.weight = 1, drop.similar.markers = TRUE, re.est.map = TRUE, ...)
}
\arguments{
\item{cross}{The qtl cross object to search}

\item{chr}{The chromosome to scan. Can be a vector of chromosome names or a single name.}

\item{rf.threshold}{The recombination fraction threshold to drop a marker. If est.rf has
not been run on cross, it will be done so automatically. See qtl::est.rf for details}

\item{sd.weight}{The weighting of segregation distortion rank in dropping a marker.
Higher values relative to na.weight increase the weight of the sd rank. Setting a value
of 0 removes sd as a factor in choosing the best marker.}

\item{na.weight}{Same as sd.weight, but for the number of nas.}

\item{drop.similar.markers}{Should the similar markers be dropped and a subsetted cross
be returned?}

\item{re.est.map}{Should the map be re-estimated after drop.similar.markers?}

\item{...}{Additional arguments passed to est.map if re.est.map = TRUE}
}
\value{
Either a character vector of the markers that should be dropped (if drop.similar.markers = FALSE),
or a new cross object with the similar markers dropped.
}
\description{
\code{findSimilarMarkers} Find markers that have a small recombination fraction and
drop the one with combined greater segregation distortion and/or missing data. ***Note:
if the cross object has many markers (>1000), avoid running this function on more than one
chromosome at a time. Also make sure to run est.rf first and use re.est.map = FALSE***
}

