% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno2cross.R
\name{geno2cross}
\alias{geno2cross}
\title{Format a genotype matrix to a R/qtl readable file}
\usage{
geno2cross(genomat, chr = NULL, pos = NULL, id = NULL, chrpos.sep = "_",
  crossfile = "cross.csv")
}
\arguments{
\item{genomat}{A matrix of individual (row) * marker (col) genotype calls.}

\item{chr}{A vector of chromosome IDs. If null, taken from the column names}

\item{pos}{A vector of marker positions. Must be able to be coerced to
numeric (does not contain letters / special characters).
If null, taken from the column names .}

\item{id}{Individual identifiers. If NULL, taken from row names of genomat.}

\item{chrpos.sep}{If chr and pos are NULL, use this character to parse
marker names into chromosome and position information. For example, if
chrpos.sep = "_", a marker named 01_100000 would be parsed into chr = 01
pos = 100000.}

\item{crossfile}{The name of the cross file to be written. Defaults to store
in current working directory. Specify full path if you want to write to another
location.}
}
\value{
Nothing - writes a file called cross.file
}
\description{
\code{geno2cross} Combine marker information and genotype data into
a R/qtl "csv" file.
}
\details{
More here
}
\examples{
\dontrun{
... more here ...
}
}
