% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoprob2marker.R
\name{genoprob2marker}
\alias{genoprob2marker}
\title{Get allelic/genotype probabilities from QTL models}
\usage{
genoprob2marker(cross, useGrid = TRUE, prob.thresh = 0, chr = NULL,
  qtl = NULL, returnNumeric = TRUE, genoNum = NULL)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{useGrid}{Should the genotype probability matrix be reduced to a grid?}

\item{prob.thresh}{What should the probability of an allele need to be to be called?}

\item{chr}{Should the analysis be limited to specific chromosomes.}

\item{qtl}{If extracting data from a single qtl model, supply here.}

\item{returnNumeric}{Should genotype calls or probabilities be returned?}

\item{genoNum}{The numeric encoding for each genotype.}

\item{...}{additional arguments passed on to rrBLUP::mixed.solve}
}
\description{
\code{genoprob2marker} Not intended for external use.
}

