% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highContrastColors.R
\name{highContrastColors}
\alias{highContrastColors}
\title{Discrete color palette (n = 256)}
\usage{
highContrastColors(n)
}
\arguments{
\item{n}{The number of unique colors to return}
}
\description{
\code{highContrastColors}  Database of 256 high contrast colors
}
\examples{
n=100
plot(1:n,runif(n), col=highContrastColors(n), pch=19)

}

