% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferMarkerPos.R
\name{inferMarkerPos}
\alias{inferMarkerPos}
\title{Find the best mapping position for a marker}
\usage{
inferMarkerPos(cross, marker.matrix, initial.geno.density = 20,
  final.step = 0.1, est.ci = FALSE, verbose = T, jitterMars = T, ...)
}
\arguments{
\item{cross}{The QTL cross object.}

\item{marker.matrix}{A numeric genotype matrix where the column names are markers and
the row names are identical to the line IDs in the cross object}

\item{initial.geno.density}{The initial distance between markers. Higher values speeds up
the initial scan for the best chromosome. Corresponds to the min.distance argument
of repPickMarkerSubset.}

\item{final.step}{The density of pseudomarkers on the final scan. Corresponds to the
step argument of calc.genoprob. Smaller values get more precision on estimates of
marker position, but slow down the algorithm.}

\item{est.ci}{Logical, should confidence intervals around the inferred position be
estimated?}

\item{...}{Additional arguments passed to scanone. Cannot be "chr" or "method" as
these are hard coded in the function.}
}
\value{
A data frame with the most likely position of the marker.
}
\description{
\code{inferMarkerPos} Used to orient scaffolds or find markers that can fill
gaps in a map.
}
\details{
There are many methods to find the most likely position of a new marker.
This approach attempts to speed up detection of the most likely position of a marker
by using the haley-knott qtl scan to find the strongest correlation between markers
in the genetic map and the numeric-coded genotypes of a new marker.
This function can be used independently to order scaffolds during genome assembly,
or to improve a genetic map.
In the latter case, the results can be piped into fillGapsInMap to find the markers
that will do the best job of increasing marker density. Once these markers are known
it is best to go back and reconstruct the genetic map. However, if only a few markers
are to be added, one could use qtl::addmarkers.
}
\examples{
library(qtlTools)
data(fake.bc)
cross<-fake.bc
\dontrun{
... more here ...
}
}
