% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans4qtl.R
\name{lsmeans4qtl}
\alias{lsmeans4qtl}
\title{Using a QTL model, calcuate conditional QTL genotype effects}
\usage{
lsmeans4qtl(cross, phe, form, mod, covar)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{phe}{Character or numeric vector indicating the phenotype to be tested}

\item{form}{Character formula that must match the QTL in the fitted QTL (mod) object.}

\item{mod}{A QTL model, generated by either makeqtl (etc.) or stepwiseqtl.}

\item{covar}{dataframe of covariates with names that match terms in the formula}
}
\value{
A dataframe of statistics.
}
\description{
\code{lsmeans4qtl} Takes a QTL model, formula and returns SAS - style LSMeans.
*** Note: to use this function, the "lsmeans" package must be installed and loaded.
}
\details{
This function iterates through the terms in the formula, pulling out lsmeans.
}

