% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTrans.R
\name{parseTrans}
\alias{parseTrans}
\title{Find the location of a cis eQTL}
\usage{
parseTrans(cross, phe, sdrop, formula = NULL, qtl, cisQTL = NULL,
  covar = NULL, cisWindow = 40, verbose = FALSE, retainCis = TRUE)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{phe}{Character or numeric vector indicating the phenotype to be tested}

\item{chr.gene}{The name of the chromosome that contains the focal genes}

\item{pos.gene}{The suspected mapping position of the gene - may be inferred from bp2cm}

\item{cmWindow}{The distance from pos.gene that the cis eQTL may reside. Default is 10, but
if the confidence of the mapping position is high, a smaller value is suggested.}

\item{outputType}{Whether to return a data.frame or qtl object (from makeqtl)}

\item{lodThreshold}{If outputType = data.frame, the presence of a cis.eqtl is defined by whether
any marker within cmWindow has a lod score > lodThreshold.}

\item{lodThreshold}{}
}
\value{
Either a qtl object or a dataframe, indicating the position of a cis eQTL.
}
\description{
\code{parseTrans} Find the mapping position of the highest lod score
within a window surrounding the known position of a gene.
}

