% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneModel.R
\name{plotGeneModel}
\alias{plotGeneModel}
\title{A plotting method for gene models}
\usage{
plotGeneModel(gff, snpEffVCF, geneID, upstreamBuffer = 1000,
  downstreamBuffer = 500, features2plot = c("exon", "five_prime_UTR",
  "three_prime_UTR"), colors = c("steelblue3", "lightsteelblue1",
  "lightsteelblue1"), mutations2annotate = c("missense", "stop", "start"),
  pchMutation = c(2, 8, 8), colMutation = c("darkblue", "darkred", "green"),
  orientation = NULL, windowSize = 100, stepSize = 10, scaleBar = 0.1,
  makePlot = TRUE, returnData = TRUE)
}
\arguments{
\item{gff}{A gene annotation in gff / gff3 format}

\item{snpEffVCF}{A simple vcf file that has been annotated by the SNPeff program}

\item{geneID}{The gff / vcf id of the gene of interest}

\item{upstreamBuffer}{How far upstream of the gene should be plotted / extracted?}

\item{downstreamBuffer}{How far downstream of the gene should be plotted / extracted?}

\item{features2plot}{What features (in the gff) should be plotted?}

\item{colors}{What colors should the features be?}

\item{mutations2annotate}{What types of mutations should be annoted in the plot?
These keywords must appear in the "INFO" column of the vcf file}

\item{pchMutation}{What shape should the annotations be?}

\item{colMutation}{What color should the annotations be?}

\item{orientation}{What is the orientation of the gene? If NULL, determined from the
gff.}

\item{windowSize}{What window size should be used in the mutation sliding window plot?}

\item{stepSize}{What step size should be used in the mutation sliding window plot?}

\item{scaleBar}{Should a scale bar be plotted?}
}
\value{
A data.frame containing all mutations in the plot.
}
\description{
\code{plotGeneModel} Using an annotated vcf and gff file, plot the gene model with
mutations.
}

