% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenicQTL.R
\name{polygenicQTL}
\alias{polygenicQTL}
\title{Run rrBLUP polygenic inheritance on an R/qtl model}
\usage{
polygenicQTL(cross, qtl = NULL, pheno.col = 1, prob.thresh = 0,
  covar = NULL, formula = NULL, rrBLUP.method = "REML",
  kinship.matrix = NULL, verbose = TRUE, ...)
}
\arguments{
\item{cross}{A qtl cross object containing conditional genotype probabilites calculated with
stepwidth = "fixed".}

\item{qtl}{A QTL model, generated by  makeqtl (etc.) .}

\item{pheno.col}{Character or numeric vector of length 1,
indicating the phenotype to be tested.}

\item{prob.thresh}{What should the probability of an allele need to be to be called?}

\item{covar}{the covariate specified in the model / formula. See fitqtl.}

\item{formula}{The formula for the qtl, names of columns in the covariate dataframe
and altnames of the QTL need to match exactly, or will be ignored.}

\item{rrBLUP.method}{should reml or ml be run?}

\item{kinship.matrix}{A kinship matrix as supplied by rrBLUP::A.mat, or similar.
If not supplied, kinship matrix is calculated on the pseudomarker grid in the cross object}

\item{verbose}{Should a normal or binary model be fit?}

\item{...}{additional arguments passed on to rrBLUP::mixed.solve}
}
\value{
A dataframe with 3 columns:
\itemize{
 \item{"Vg"}{The genetic variance attributable to the kinship (polygenic)}
 \item{"Ve"}{The amount of environmental (residual) varaince}
 \item{"polyH2"}{the proportion of Vtotal in Vg. Equivalent to the amount of
 heritability attributable to kinship (polygenic) variance following incoporpation of
 any fixed effects}
}
}
\description{
\code{polygenicQTL} Parses and pushes qtl model data int rrBLUP's mixed model
genomic selection pipe. Outputs polygenic inheritance stats.
}
\details{
See rrBLUP::mixed.solve and rrBLUP::kin.blup
}
\examples{
\dontrun{
library(qtlTools)
library(rrBLUP)
data(fake.bc)
cross<-fake.f2
cross<-calc.genoprob(cross, step = 12, stepwidth = "fixed")
qtl<-makeqtl(cross, chr = 13, pos = 28, what = "prob")
formula = " ~ Q1 + Q1*sex + sex"
covar = data.frame(sex = pull.pheno(cross, pheno.col = "sex"))
polygenicQTL(cross = cross, pheno.col = 1)
polygenicQTL(cross = cross, covar = covar, formula = "~ sex", pheno.col = 1)
polygenicQTL(cross = cross, qtl = qtl, formula = "~ Q1", pheno.col = 1)
polygenicQTL(cross = cross, qtl = qtl,covar = covar, formula = "~ Q1 + Q1*sex + sex", pheno.col = 1)
}

}

