% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullMap.R
\name{pullMap}
\alias{pullMap}
\title{Pull the genetic map and marker names}
\usage{
pullMap(cross)
}
\arguments{
\item{cross}{The qtl cross object.}
}
\description{
\code{pullMap} An extension of qtl::pull.map
}

