% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullSigQTL.R
\name{pullSigQTL}
\alias{pullSigQTL}
\title{Method to summarize scanone results}
\usage{
pullSigQTL(cross, s1.output, perm.output, pheno.col = NULL, chr = NULL,
  alpha = 0.05, returnQTLModel = TRUE, ...)
}
\arguments{
\item{cross}{The qtl cross.}

\item{s1.output}{The output from scanone}

\item{perm.output}{The permutation output from scanone}

\item{pheno.col}{Character or numeric vector indicating the phenotype to be tested.}

\item{chr}{The chromosome to be tested. Defaults to all chromosomes.}

\item{alpha}{The significance for permutations}

\item{returnQTLModel}{Logical, should a QTL model be returned (TRUE), or
should a culled output from qtlpvl::convert_scan1 be returned (FALSE)?}

\item{...}{additional arguments passed on to summary.scanone, such as controlAcrossCol.}
}
\value{
Either QTL models or simplified and converted scanone summary.
}
\description{
\code{pullSigQTL} Uses qtlpvl to summarize the output of scanone,
then culls the output to only significant QTL peaks, based on permutations.
}
\examples{
library(qtlTools)
data(fake.bc)
cross<-fake.bc
cross <- calc.genoprob(cross, step=2.5)
s1<-scanone(cross, method="hk", pheno.col=c("pheno1", "pheno2"))
perm<-scanone(cross, n.perm=100, method="hk",pheno.col=c("pheno1", "pheno2"), verbose=F)
pullSigQTL(cross, s1.output=s1, perm.output=perm)
pullSigQTL(cross, s1.output=s1, perm.output=perm, returnQTLModel=F)

}

