% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlSignTest.R
\name{qtlSignTest}
\alias{qtlSignTest}
\title{Conduct the sign test for expression bias}
\usage{
qtlSignTest(effect = NULL, trans.effect = NULL, verbose = T, ...)
}
\arguments{
\item{effect}{Signed cis-eQTL or allele-specific expression log2-fold change data}

\item{trans.effect}{If effect is signed cis-eQTL effect, one can test the independence
of cis and trans eQTL evolution. Here provide effect effect of trans QTL.}

\item{verbose}{should results be printed to console?}

\item{...}{Additional arguments to pass on to fisher.test / binom.test, such as
null hypotheses.}

\item{id}{label for output}
}
\value{
The results of a binomial test (or Fisher's exact test if trans.dir is specified).
}
\description{
\code{qtlSignTest} Uses the Fraser etc. sign test to assess neutral evolution
for a set of expression phenotypes
}
\details{
See ...
}

