% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantnorm.R
\name{quantnorm}
\alias{quantnorm}
\title{Quantile normalize vector of numeric data}
\usage{
quantnorm(x, drop0s = FALSE)
}
\arguments{
\item{x}{The numeric vector to quantile normalize}
}
\value{
a vector of quantile normalized data
}
\description{
\code{quantnorm} Rank-normalize (quantile normalize) data to enforce strict normality
of data. Useful if assuming identical distributions across many response variables
}
\examples{
par(mfrow = c(2,1))
hist(x <- sample(1:100, 50), breaks = 12)
hist(quantnorm(x), breaks = 12)

library(qtl)
data(fake.f2)
hist(pull.pheno(fake.f2, "phenotype"))
hist(quantnorm(pull.pheno(fake.f2, "phenotype")))
fake.f2<-transformPheno(fake.f2, pheno.col = "phenotype", transf = quantnorm)
hist(pull.pheno(fake.f2, "phenotype"))

}

