% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reDoCross.R
\name{reDoCross}
\alias{reDoCross}
\title{Run a set of functions that improve the genetic map}
\usage{
reDoCross(cross, window = 5, min.distance = 1, map.function = "kosambi",
  sex.sp = F, verbose = T, initialEstMap = TRUE, ...)
}
\arguments{
\item{cross}{The QTL cross object.}

\item{window}{Passed to ripple. Smaller values (e.g. 3) run more quickly but explore
fewer possible marker orders. Larger values (e.g. 6) should only be used on maps with
few markers. Large numbers of markers with large windows will take a long time.}

\item{min.distance}{The minimum distance between markers in the final map. If two markers
are closer than this value, drop the one with fewer NAs and better segregation distortion.}

\item{map.function}{The map function to pass on to est.map}

\item{sex.sp}{Should sex-specific maps be estimated in a 4-way cross? Passed on to est.map.}

\item{verbose}{Should updates be printed ot the terminal?}

\item{initialEstMap}{Should the map be estimated before any functions are run?}
}
\value{
A cross object with markers reordered.
}
\description{
\code{reDoCross} Drop close markers, re-order markers via ripple, then re-check
for close markers. Returns a cross object with improved marker order and overly
dense marker regions removed.
}
\details{
If any chromosomes are sex-specific, re-class such chromosomes after running
newLG: using class(cross$geno[["Xchrom"]]) <- "S" or whatever. The new map has
arbitrary cM positions where each marker is separated by 10cM. Run est.map to
get true cM positions.
}
\examples{
library(qtlTools)
data(fake.f2)
cross<-fake.f2
\dontrun{
map=est.map(cross, map.function = "haldane")
cross=replace.map(cross, map)
cross2<-reDoCross(cross, map.function = "haldane")
plot.map(cross, cross2)
}

}
