% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameMarkers.R
\name{renameMarkers}
\alias{renameMarkers}
\title{Change marker names}
\usage{
renameMarkers(cross, crossfile, oldnames, newnames, outputName)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{crossfile}{The location of the .csv file containing the genotype matrix. Formats accepted
are only the .csv and .csvs r/qtl read.cross formats. Other formats may run without error, but
do not work.}

\item{oldnames}{A character vector of the names of markers that need to be changed}

\item{newnames}{A character vector of the names of markers to replace the current ones}

\item{outputName}{A character string with the name of the new genotype matrix with new names}
}
\value{
A genotype matrix is written to file in the current working directory, unless a path is
specificed in the "crossfile" and "outputName" arguments. The new matrix will have marker names
that have been replaced by the newnames provided.
}
\description{
\code{renameMarkers} Take a cross object and the corresponding genotype matrix
and rename a set of specified markers.
}
\examples{
library(qtl)
data(fake.f2)
write.cross(fake.f2, format = "csv", filestem = "test") ### Must be "csv" or "csvs" formats.
old.x <- markernames(fake.f2)[grep("X", markernames(fake.f2))]
new.x <- c("X1", "X2", "X3")
renameMarkers(cross = fake.f2, crossfile = "test.csv",
   oldnames = old.x, newnames = new.x, outputName = "new.fake.f2.csv")
new.fake.f2 <- read.cross(file = "new.fake.f2.csv", format = "csv")
cbind(markernames(fake.f2), markernames(new.fake.f2))
}

