% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatRipple.R
\name{repeatRipple}
\alias{repeatRipple}
\title{Method to improve a genetic map.}
\usage{
repeatRipple(cross, chr, window = 6, repeatloop = TRUE, makePlots = FALSE,
  method = "countxo", ...)
}
\arguments{
\item{cross}{The qtl cross object to search}

\item{chr}{The chromosome to scan. Can be a vector of chromosome names or a single name.}

\item{window}{The window size passed on to ripple}

\item{repeatloop}{Should ripple be run once, or iteratively?}

\item{makePlots}{Should the recombination fraction be iteratively plotted?}

\item{method}{The method passed on to ripple}

\item{...}{Additional arguments passed to qtl::ripple}
}
\value{
A new cross object with the optimal rippled marker order
}
\description{
\code{repeatRipple} Iteratively run ripple.
}

