% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repPickMarkerSubset.R
\name{repPickMarkerSubset}
\alias{repPickMarkerSubset}
\title{Method to improve an estimated genetic map.}
\usage{
repPickMarkerSubset(cross, chr = NULL, na.weight = 2, sd.weight = 1,
  min.distance = 1, verbose = TRUE)
}
\arguments{
\item{cross}{The qtl cross object to search}

\item{chr}{The chromosome to scan. Can be a vector of chromosome names or a single name.
If NULL, run on all chromosomes.}

\item{na.weight}{Same as sd.weight, but for the number of NAs.
Only used if cross is not a 4way.}

\item{sd.weight}{The weighting of segregation distortion rank in dropping a marker.
Higher values relative to na.weight increase the weight of the sd rank. Setting a value
of 0 removes sd as a factor in choosing the best marker. Only used if cross is not a 4way.}

\item{min.distance}{Logical, should markers on the ends of the chromosomes always be retained?}

\item{verbose}{Logical, should updates be printed?}
}
\value{
A new cross object with the similar markers dropped.
}
\description{
\code{repPickMarkerSubset} finds the best marker within a cM window.
}
\examples{
set.seed(42)
library(qtlTools)
map<-sim.map(len = c(50,20), n.mar = c(20,30), include.x=FALSE)
cross0<-sim.cross(map, n.ind=50, type="f2", map.function="kosambi",
   error.prob=.01, missing.prob = .05)
cross0<-est.rf(cross0)
cross1<-repPickMarkerSubset(cross0)
par(mfrow=c(2,1))
plot.map(cross0, cross1, main = "comparison of full and culled maps")

}
