% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repRipple.R
\name{repRipple}
\alias{repRipple}
\title{Method to improve a genetic map.}
\usage{
repRipple(cross, chr = NULL, window = 6, repeatloop = TRUE,
  makePlots = FALSE, method = "countxo", re.est.map = TRUE,
  map.function = "kosambi", error.prob = 0.001, verbose = TRUE, ...)
}
\arguments{
\item{cross}{The qtl cross object to search}

\item{chr}{The chromosome to scan. Can be a vector of chromosome names or a single name.}

\item{window}{The window size passed on to ripple}

\item{repeatloop}{Should ripple be run once, or iteratively?}

\item{makePlots}{Should the recombination fraction be iteratively plotted?}

\item{method}{The method passed on to ripple}

\item{re.est.map}{Should the map be re-estimated after drop.similar.markers?}

\item{map.function}{If re.est.map = TRUE, what map function should be used?}

\item{error.prob}{If re.est.map = TRUE, what error probability should be used?}

\item{verbose}{logical, should updates be reported?}

\item{...}{Additional arguments passed to qtl::ripple}
}
\value{
A new cross object with the optimal rippled marker order
}
\description{
\code{repRipple} iteratively runs ripple,
returning a cross object with the best order.
}
\examples{
\dontrun{
set.seed(42)
map<-sim.map(len = c(50,20), n.mar = c(20,40), include.x=FALSE)
cross0<-sim.cross(map, n.ind=50, type="f2", map.function="kosambi")
## Mess up the order on chr1
mars<-1:nmar(cross0)[1]
set.seed(42)
badorder<-order(jitter(mars, factor = 10))
cross0<-switch.order(cross0, chr = 1, order = badorder,
                     error.prob=0.001, map.function="kosambi")
newmap<-est.map(cross0, error.prob=0.001, map.function="kosambi")
cross0 <- replace.map(cross0, newmap)
cross0<-est.rf(cross0)
cross1<-repRipple(cross0, chr = 1, error.prob=0.001, map.function="kosambi",window = 6)
plot.rf(cross0, chr = 1, main = "recombination fractions before ripple")
plot.rf(cross1, chr = 1, main = "recombination fractions after ripple")
}
}

