% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan4trans.R
\name{scan4trans}
\alias{scan4trans}
\title{Find the location of a cis eQTL}
\usage{
scan4trans(cross, phe, perms = NULL, nperms = 1000, k = 0,
  cisQTL = NULL, cisGenotypeCovariate = TRUE, covar = NULL,
  cisWindow = 40, verbose = FALSE, cisname = "Q1", refine = TRUE, ...)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{phe}{Character or numeric vector indicating the phenotype to be tested}

\item{chr.gene}{The name of the chromosome that contains the focal genes}

\item{pos.gene}{The suspected mapping position of the gene - may be inferred from bp2cm}

\item{cmWindow}{The distance from pos.gene that the cis eQTL may reside. Default is 10, but
if the confidence of the mapping position is high, a smaller value is suggested.}

\item{outputType}{Whether to return a data.frame or qtl object (from makeqtl)}

\item{lodThreshold}{If outputType = data.frame, the presence of a cis.eqtl is defined by whether
any marker within cmWindow has a lod score > lodThreshold.}

\item{lodThreshold}{}
}
\value{
Either a qtl object or a dataframe, indicating the position of a cis eQTL.
}
\description{
\code{scan4trans} Find the mapping position of the highest lod score
within a window surrounding the known position of a gene.
}

