% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scantwo.GWERk.R
\name{scantwo.GWERk}
\alias{scantwo.GWERk}
\title{Scanone qtl analysis permitting GWERk analysis.}
\usage{
scantwo.GWERk(cross, n.perm = 100, ..., gwerk = 1,
  use.scantwopermhk = T)
}
\description{
\code{scantwo.GWERk}  see qtl::scanone for details. GWERk = 0 runs normal scanone
permutations. Otherwise GWERk perms are used.
}
