% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scantwo.GWERk1.R
\name{scantwo.GWERk1}
\alias{scantwo.GWERk1}
\title{Scantwo qtl analysis forcing GWERk = 1 analysis.}
\usage{
scantwo.GWERk1(cross, chr, pheno.col = 1, model = c("normal", "binary"),
  method = c("em", "imp", "hk", "mr", "mr-imp", "mr-argmax"),
  addcovar = NULL, intcovar = NULL, weights = NULL, use = c("all.obs",
  "complete.obs"), incl.markers = FALSE, clean.output = FALSE,
  clean.nmar = 1, clean.distance = 0, maxit = 4000, tol = 1e-04,
  verbose = TRUE, n.perm, perm.Xsp = FALSE, perm.strata = NULL,
  assumeCondIndep = FALSE, batchsize = 250, n.cluster = 1)
}
\description{
\code{scanone.GWERk}  see qtl::scantwo for details. Forces GWERk = 1.
}

