% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsOnMap.R
\name{segmentsOnMap}
\alias{segmentsOnMap}
\title{Visualize the positions of QTL on a genetic map.}
\usage{
segmentsOnMap(cross, phe, chr, l, h, segSpread = 0.15, ...)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{phe}{Character or numeric vector indicating the phenotype to be tested}

\item{chr}{Vector of chromosome ids - will be coerced to numeric}

\item{l}{The lower confidence interval bound for each qtl}

\item{h}{The upper confidence interval bound for each qtl}

\item{...}{Other arguments passed to segments}

\item{seqSpread}{How far apart (x axis) the semgents are}
}
\value{
The plot
}
\description{
\code{segmentsOnMap} A basic function to plot QTL confidence intervals.
Useful for <10 traits. All inputs (excluding cross) must have the same length,
since each element represents a single segment on the map.
}
\details{
Pass output from bayesint, lodint, or another confidence
interval estimation program to visualize this.
}

