% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitText.R
\name{splitText}
\alias{splitText}
\title{Simplify strsplit functionality}
\usage{
splitText(x, sep = "_", num = 1)
}
\arguments{
\item{x}{A character vector, if not character, function coerces it to character.}

\item{sep}{The character that separates values}

\item{num}{The index of the split vector to return.}
}
\description{
\code{splitText} Loops along a string splitting text by sep and returning
the element num of that split text
}
