% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jtlbar.R
\name{theme_jtlbar}
\alias{theme_jtl}
\alias{theme_jtlbar}
\title{simple custom themes}
\usage{
theme_jtlbar()
}
\description{
\code{theme_jtlbar}  two themes for ggplot similar to theme bw, but better (i think). One is
appropriate for most types of plots (theme_jtl). The other is better for barplots (theme_jtlbar)
}

