% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covScanQTL.R
\name{covScanQTL}
\alias{covScanQTL}
\title{Find the effect of candidate genes based on gene expression.}
\usage{
covScanQTL(cross, pheno.y, pheno.candidates, experimental.covar = NULL,
  mod = NULL, QTLxE = TRUE, chr, pos, qtl.method = "hk", nperm = 0,
  plotit = TRUE)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{pheno.y}{Character vector specifying the phenotypes modelled. Is passed
to the "pheno.col" argument of qtl::scanone.}

\item{pheno.candidates}{Character vector specifying the names of the expression phenotypes to
be extracted from the cross object that serve as the covariate.}

\item{mod}{A qtl model object that can supply additional marker covariates that
are held constant during the covariate scan.}

\item{QTLxE}{Logical, should the focal QTL have an interactive covariate.
See exp.covar.}

\item{chr}{Numeric or character, of length 1, specifying the chromosome on which
the QTL peak to scan exists}

\item{pos}{Numeric, of length 1, specifying the QTL peak position on chromosome chr.}

\item{qtl.method}{The method passed to scanone}

\item{nperm}{If permutation tests are desired, specify as >0. These are not currently
computationally efficient and take forever when using eQTL data.}

\item{plotit}{Logical, when more than 1 pheno.y is specified, presents a boxplot of
covariate scan ranks.}

\item{exp.covar}{data.frame containing the experimental covariate. Passed on to
addcovar in qtl::scanone. If QTLxE is TRUE, also passed to intcovar}
}
\value{
Either a list or a dataframe, containing the maximum scanone outputs at
chromosome chr and position pos for each phenotype and covariate. If nperm > 0, also
gives the permutation results as a second list element.
}
\description{
\code{covScanQTL} Employs the covariate scan approach (Lovell et al. (2015),
Plant Cell), to rank the potential of candidate genes based on their effect
on QTL morphology. Can be run for a single phenotype (e.g. Lovell et al. (2015)),
or on a set of QTL underlying a trand-band. In the latter case, a boxplot of ranks
can be output.
}
\examples{
# See tutorial - findingCandidateGenes

}

