% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanScan.R
\name{meanScan}
\alias{meanScan}
\title{Make a scanone plot of genotype means.}
\usage{
meanScan(cross, pheno.col = 1, chr = NULL, covar = NULL,
  mean.FUN = function(x) mean(x, na.rm = T), se.FUN = function(x) sd(x,
  na.rm = T)/sqrt(sum(!is.na(x))), sw.width = 1, plot.se = T,
  se.alpha = 0.2, col = NULL, leg.pos = "topright", ...)
}
\arguments{
\item{cross}{The qtl cross. Must contain genotype probabilities or imputed genotypes}

\item{pheno.col}{Character or numeric vector indicating the phenotype to be tested.
Only 1 phenotype can be tested at a time. If numeric, converted to character.}

\item{chr}{The chromosomes to analyze. If not supplied, all are tested.}

\item{covar}{dataframe of covariates like addcovar in qtl::scanone. Can only be a single column.
If the data is numeric, or character, converted to factor first.}

\item{mean.FUN}{The function applied to generate the line positions. Defaults to mean(x, na.rm = T)}

\item{se.FUN}{The function applied to generate the confidence interval around the mean.
Defaults to sd(x, na.rm = T)/sqrt(sum(!is.na(x)))}

\item{sw.width}{The window size to use if smoothing is desired. If 1, no sliding window smoothing
is conducted. Must be smaller than the number of markers on the smallest chromosome.}

\item{plot.se}{Logical, should standard errors around the menas be plotted?}

\item{se.alpha}{Numeric coding of the transparency for the standard error regions.}

\item{col}{Character vector of colors to plot.}

\item{leg.pos}{The character of numeric (length 2) position of the legend. Default is "topright".}

\item{...}{additional arguments passed on to plot.scanone}

\item{leg.pos}{The position of the legend.}
}
\value{
The plot and an object of class scanone containing the means
for each combination of genotype and covariate.
}
\description{
\code{meanScan} Like qtl:effectscan, but plots the genotypic means across the
marker / pseudomarker grid.
}
\details{
Calculates genotypic means at each marker/psuedomarker, plots them.
}
\examples{
\dontrun{
library(qtlTools)
library(qtl)
data(fake.bc)
cross<-fake.bc
cross<-calc.genoprob(cross)
ms<-meanScan(cross, covar = NULL)

# Some simple customization
ms<-meanScan(cross, covar = NULL, col = c("black","orange"),
   leg.pos = "top")

sex<-data.frame(sex = ifelse(pull.pheno(cross, "sex") == 0,"F","M"))
ms<-meanScan(cross, covar = sex, leg.pos = "right")
ms<-meanScan(cross, covar = sex, chr = c(2,7), leg.pos = "right",
   col = c("black","grey","darkblue","lightblue"))

cross.fem<-subset(cross, ind = sex[,1] == "F")
cross.male<-subset(cross, ind = sex[,1] == "M")
par(mfrow = c(2,1))
ms <-meanScan(cross.fem, leg.pos = "topright", leg.inset = .05, main = "females only")
ms <-meanScan(cross.male, leg.pos = "topright", leg.inset = .05, main = "males only")

data(fake.f2)
cross<-fake.f2
cross<-calc.genoprob(cross)
ms <- meanScan(cross, covar = NULL, leg.pos = "bottomright", leg.inset = .1)
sex<-data.frame(sex = ifelse(pull.pheno(cross, "sex") == 0,"F","M"))
ms<-meanScan(cross, covar = sex)

data(fake.4way)
cross<-fake.4way
cross<-calc.genoprob(cross)
ms<-meanScan(cross, covar = NULL)
sex<-data.frame(sex = ifelse(pull.pheno(cross, "sex") == 0,"F","M"))
ms<-meanScan(cross, covar = sex, chr = c(2,7))

data("multitrait")
cross =multitrait
cross<-calc.genoprob(cross)
ms<-meanScan(cross, covar = NULL)
ms<-meanScan(cross, covar = NULL, sw.width = 5)
}

}
