% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantnorm.R
\name{quantnorm}
\alias{quantnorm}
\title{Quantile normalize vector of numeric data}
\usage{
quantnorm(x, drop0s = FALSE)
}
\arguments{
\item{x}{The numeric vector to quantile normalize}

\item{drop0s}{Logical, should 0's be transformed to NAs}
}
\value{
a vector of quantile normalized data
}
\description{
\code{quantnorm} Rank-normalize (quantile normalize) data to enforce strict normality
of data. Useful if assuming identical distributions across many response variables
}
\examples{
library(qtlTools)
data(fake.bc)
cross<-fake.bc
par(mfrow = c(2,1))
hist(pull.pheno(cross, pheno.col = "pheno2"), breaks=20,
   main = "before quantile normalization",
   xlab = "raw pheno1")
cross<-transformPheno(cross, pheno.col = "pheno1", quantnorm)
hist(pull.pheno(cross, pheno.col = "pheno1"), breaks=20,
   main = "after quantile normalization",
   xlab = "qn pheno1")

}
