% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantnorm.R
\name{quantnorm}
\alias{quantnorm}
\title{Quantile normalize vector of numeric data}
\usage{
quantnorm(x, drop0s = FALSE)
}
\arguments{
\item{x}{The numeric vector to quantile normalize}
}
\value{
a vector of quantile normalized data
}
\description{
\code{quantnorm} Rank-normalize (quantile normalize) data to enforce strict normality
of data. Useful if assuming identical distributions across many response variables
}

