% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repRipple.R
\name{repRipple}
\alias{repRipple}
\title{Method to improve a genetic map.}
\usage{
repRipple(cross, chr = NULL, window = 6, repeatloop = TRUE,
  makePlots = FALSE, method = "countxo", re.est.map = TRUE,
  map.function = "kosambi", error.prob = 0.001, verbose = T, ...)
}
\arguments{
\item{cross}{The qtl cross object to search}

\item{chr}{The chromosome to scan. Can be a vector of chromosome names or a single name.}

\item{window}{The window size passed on to ripple}

\item{repeatloop}{Should ripple be run once, or iteratively?}

\item{makePlots}{Should the recombination fraction be iteratively plotted?}

\item{method}{The method passed on to ripple}

\item{re.est.map}{Should the map be re-estimated after drop.similar.markers?}

\item{map.function}{If re.est.map = TRUE, what map function should be used?}

\item{error.prob}{If re.est.map = TRUE, what error probability should be used?}

\item{...}{Additional arguments passed to qtl::ripple}
}
\value{
A new cross object with the optimal rippled marker order
}
\description{
\code{repRipple} Iteratively run ripple.
}

