% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsOnMap.R
\name{segmentsOnMap}
\alias{segmentsOnMap}
\title{Visualize the positions of QTL on a genetic map.}
\usage{
segmentsOnMap(cross, phe, chr, l, h, peaklod = NA, calcCisResults = NULL,
  legendPosition = "bottom", legendCex = 0.8, col = NULL,
  palette = highContrastColors, lwd = "byLod", leg.lwd = 2, max.lwd = 5,
  min.lwd = 1, tick.width = NULL, leg.inset = 0.01, chrBuffer = c(0.05,
  0.15), orderBy = "lod", ...)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{phe}{Character or numeric vector indicating the phenotype to be tested}

\item{chr}{Vector of chromosome ids - will be coerced to numeric}

\item{l}{The lower confidence interval bound for each qtl}

\item{h}{The upper confidence interval bound for each qtl}

\item{peaklod}{Optional - if provided, permits segments width
to be weighted by significance.}

\item{calcCisResults}{A shortcut that allows results from calcCis
to be piped directly into the plotting function. If provided,
l, h, chr, phe and peaklod are ignored}

\item{legendPosition}{Where to place the legend. This is passed
to the first argument of the legend function. See legend documentation}

\item{legendCex}{The character expansion for the legend. This is
passed to the cex argument of legend.}

\item{col}{Optional. A vector of colors for segments}

\item{palette}{Optional. A color palette with which to draw segment colors}

\item{lwd}{Either the specification "byLod", where segment weights
are calculated from LOD score data, or a numeric vector of length
1 that is passed to segments lwd}

\item{leg.lwd}{The lineweights of the legend.}

\item{max.lwd}{If scaling by LOD, what should the maximum lwd be?}

\item{min.lwd}{If scaling by LOD, what should the minimum lwd be?}

\item{tick.width}{How wide should the ticks be? Defaults to 1/n chromosomes.}

\item{leg.inset}{How far should the legend be away from the plot border?}

\item{chrBuffer}{Numeric vector of length 2, specifying the buffer space
between confidence interval segments on the left and right adjacent chromosomes.}

\item{...}{Other arguments passed to segments}
}
\value{
The plot
}
\description{
\code{segmentsOnMap} A basic function to plot QTL confidence intervals.
Useful for <20 traits.
}
\details{
Pass output from bayesint, lodint, or another confidence
interval estimation program to visualize this.
}
\examples{

library(qtlTools)
data(multitrait)
segmentsOnMap(cross = multitrait, phe = paste0("phe",1:15),
 chr = rep(1,15), l =seq(from = 1, to = 100, length.out = 15),
 h =seq(from = 20, to = 120, length.out = 15)
segmentsOnMap(cross = multitrait, phe = paste0("phe",1:25),
 chr = rep(1,25), l =seq(from = 1, to = 100, length.out = 25),
 h =seq(from = 20, to = 120, length.out = 25))
\dontrun{
# use multitrait data
cross <- multitrait
cross <- calc.genoprob(cross)
phes <- phenames(cross)[1:20]
# Conduct scanone and permutations to get confidence intervals
s1 <- scanone(cross, pheno.col = phes, method = "hk")
perms <- scanone(cross, pheno.col = phes, method = "hk",
   n.perm = 100, verbose=F)
cis<-calcCis(s1.output = s1, perm.output = perms)

# manual construction of the confidence intervals
with(cis, segmentsOnMap(cross, phe = pheno, chr = chr,
   l = lowposition, h = highposition, legendCex = .5,
   tick.width = .1,  chrBuffer = c(.15,.2)))

# feed calcCis directly into the plot
segmentsOnMap(cross, calcCisResults = cis, legendCex = .5)
segmentsOnMap(cross, calcCisResults = cis, legendCex = .5,
   col = terrain.colors(length(unique(cis$phe))))
segmentsOnMap(cross, calcCisResults = cis, legendCex = .5,
   max.lwd=6, min.lwd=.5)
segmentsOnMap(cross, calcCisResults = cis, legendCex = .5,
   max.lwd=4, min.lwd=2)
segmentsOnMap(cross, calcCisResults = cis, legendCex = .5,
   lwd = 2)
segmentsOnMap(cross, calcCisResults = cis, legendCex = .5,
   palette = rainbow)
}
}

