% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsOnMap.R
\name{segmentsOnMap}
\alias{segmentsOnMap}
\title{Visualize the positions of QTL on a genetic map.}
\usage{
segmentsOnMap(cross, phe, chr, l, h, segSpread = 0.15,
  legendPosition = "bottom", legendCex = 1, chr.adj = NULL,
  jColors = NULL, ...)
}
\arguments{
\item{cross}{The qtl cross object with marker names that need to be changed.}

\item{phe}{Character or numeric vector indicating the phenotype to be tested}

\item{chr}{Vector of chromosome ids - will be coerced to numeric}

\item{l}{The lower confidence interval bound for each qtl}

\item{h}{The upper confidence interval bound for each qtl}

\item{legendPosition}{Where to place the legend. This is passed to the first argument of
the legend function. See legend documentation}

\item{legendCex}{The character expansion for the legend. This is passed to the cex argument
of legend.}

\item{chr.adj}{How far should the first segment be from the map drawing? Defaults to scale with the
number of chromosomes}

\item{jColors}{A dataframe with two columns, "phe" and "color". The phe column contains a all unique
phenotypes and the color is the associated color. The default sets the colors to the rainbow color
palette.}

\item{...}{Other arguments passed to segments}

\item{seqSpread}{How far apart (x axis) the semgents are}
}
\value{
The plot
}
\description{
\code{segmentsOnMap} A basic function to plot QTL confidence intervals.
Useful for <10 traits. All inputs (excluding cross) must have the same length,
since each element represents a single segment on the map.
}
\details{
Pass output from bayesint, lodint, or another confidence
interval estimation program to visualize this.
}
\examples{
# See basicQTL tutorial

}

