% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentsOnPeaks.R
\name{segmentsOnPeaks}
\alias{segmentsOnPeaks}
\title{Single QTL confidence intervals on LOD curves}
\usage{
segmentsOnPeaks(cross, calcCisOutput = NULL, ci.chr = NULL, peak = NULL,
  l = NULL, h = NULL, s1.output = NULL, mod = NULL,
  showallchr = FALSE, chr = NULL, pt.pch = 8, pt.cex = 0.8,
  pt.col = "red", int.y = 0, ...)
}
\arguments{
\item{cross}{R/qtl cross object}

\item{calcCisOutput}{The output from qtlTools::calcCis. If this is not
supplied, ci.chr, peak, l, h must be. If calcCisOutput, these arguments are
ignored}

\item{ci.chr}{vector of chromosomes for segments}

\item{peak}{vector of QTL peak positions (cM) for segments}

\item{l}{vector of QTL lower confidence interval bounds (cM) for segments}

\item{h}{vector of QTL upper confidence interval bounds (cM) for segments}

\item{s1.output}{If plotting on a scanone LOD plot, provide the scanone object
used previously.}

\item{mod}{If plotting on lodProfile, pass the qtl model object that was input into
plotLodProfile}

\item{showallchr}{Must match call to plotLodProfile if qtl is specified}

\item{chr}{If plotting on a specified set of chromosomes via scanone, replicate
the chr call here.}

\item{pt.pch}{pch parameter to points}

\item{pt.cex}{cex parameter to points}

\item{pt.col}{color parameter to points}

\item{int.y}{the y intercept of the segments}

\item{...}{additional parameters passed to segments, e.g. col, lty, etc.}
}
\description{
\code{segmentsOnPeaks} Overlay confidence intervals onto a LOD plot,
from as either plotLodProfile or plot.scanone.
}
\examples{
\dontrun{
library(qtlTools)
data(fake.bc)
cross<-fake.bc
cross <- calc.genoprob(cross, step=2.5)

# Make a QTL object and formula
mod <- makeqtl(cross, chr = c(2,5), pos = c(40,25), what = "prob")
sex <- data.frame(sex = cross$phe$sex)
nform <- "y ~ Q1 + Q2 + Q1*sex + sex"
#Calculate lodprofiles for confidence interval estimation
mod <- refineqtl(cross, mod, pheno.col = "pheno1",
                 qtl = mod, formula = nform, covar = sex, method="hk")

cis<-calcCis(mod)
plotLodProfile(mod, showallchr=FALSE)
segmentsOnPeaks(cross, qtl = mod, calcCisOutput = cis,  showallchr=FALSE)

plotLodProfile(mod, showallchr=TRUE)
segmentsOnPeaks(cross, qtl = mod, calcCisOutput = cis, showallchr=TRUE)
cis<-calcCis(mod, drop = 3)
segmentsOnPeaks(cross, qtl = mod, calcCisOutput = cis, showallchr=TRUE,
   col = "purple", int.y=.5)

s1<-scanone(cross, method="hk", pheno.col="pheno1")
perm<-scanone(cross, n.perm=100, method="hk",pheno.col="pheno1", verbose=FALSE)
cis<-calcCis(s1.output=s1, perm.output=perm)
plot(s1)
segmentsOnPeaks(cross, s1.output=s1, calcCisOutput = cis)
plot(s1, chr = c(2,5))
segmentsOnPeaks(cross, s1.output=s1, calcCisOutput = cis, chr = c(2,5))
}
}
