% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspOrder.R
\name{tspOrder}
\alias{tspOrder}
\title{Order markers within linkage groups}
\usage{
tspOrder(cross, markerList, method = "nn", hamiltonian = TRUE,
  concorde_path = NULL, return = "cross")
}
\arguments{
\item{cross}{The QTL cross object.}

\item{method}{The solve_TSP method to employ.}

\item{hamiltonian}{Logical, should a hamiltonian circuit be enforced?}

\item{concorde_path}{Required if method = "concorde". The directory containing
concorde executables.}

\item{return}{If "cross" is specified, pass the marker order through newLG, which quickly
reorders the markers of the original cross to follow the TSP order. Otherwise,
return a named list of markers in order for each chromosome.}

\item{...}{Additional arguments passed on to solve_TSP that are then paseed to TSP::contol.}
}
\value{
Either a cross object with reordered markers, or a named list
of markers in a new order.
}
\description{
\code{tspOrder} Apply a traveling salesperson problem solver to find the
shortest path through the recombination fraction matrix. Requires that
the TSP package is installed. Optimal performance is provided by the
"Concorde" method; however, this requires independent installation of
the Concorde program. See details.
}
\details{
This function relies on the TSP R packages to perform the TSP solvers. See documentation
therein, esspecially the function TSP::solve_TSP. We permit inference of marker order within a
Hamiltonian circuit by adding a dummy node that has 0 distance to all other nodes. This
allows for discrete start and end points and is more appropriate for genetic map construction
than forcing a complete route through all markers.

The best performance seems to result from using hamiltonian = T and method = "concorde".

We recommend using the concorde algorithm. To do this, install the concorde program:
http://www.math.uwaterloo.ca/tsp/concorde/downloads/codes/src/co031219.tgz.
If on a mac, this is not super easy.
See https://qmha.wordpress.com/2015/08/20/installing-concorde-on-mac-os-x/
for details on the best way to do this.
}
\examples{
library(qtlTools)
data(fake.bc)
cross<-fake.bc
\dontrun{
... more here ...
}
}

